refactor, bulk insert product move to repository
